# api/rest/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package rest

import (
	"fmt"
	"net/url"
	"time"

	"github.com/go-faster/errors"
)

func (s *ProblemStatusCode) Error() string {
	return fmt.Sprintf("code %d: %+v", s.StatusCode, s.Response)
}

// Public, unauthenticated configuration. Never put a secret here.
// Ref: #/components/schemas/Config
type Config struct {
	// The daemon's display name.
	Name string `json:"name"`
	// The build's version.
	Version string `json:"version"`
}

// GetName returns the value of Name.
func (s *Config) GetName() string {
	return s.Name
}

// GetVersion returns the value of Version.
func (s *Config) GetVersion() string {
	return s.Version
}

// SetName sets the value of Name.
func (s *Config) SetName(val string) {
	s.Name = val
}

// SetVersion sets the value of Version.
func (s *Config) SetVersion(val string) {
	s.Version = val
}

// Ref: #/components/schemas/Health
type Health struct {
	Status HealthStatus `json:"status"`
	// The build's version.
	Version string `json:"version"`
	// The daemon's clock when it answered.
	At time.Time `json:"at"`
}

// GetStatus returns the value of Status.
func (s *Health) GetStatus() HealthStatus {
	return s.Status
}

// GetVersion returns the value of Version.
func (s *Health) GetVersion() string {
	return s.Version
}

// GetAt returns the value of At.
func (s *Health) GetAt() time.Time {
	return s.At
}

// SetStatus sets the value of Status.
func (s *Health) SetStatus(val HealthStatus) {
	s.Status = val
}

// SetVersion sets the value of Version.
func (s *Health) SetVersion(val string) {
	s.Version = val
}

// SetAt sets the value of At.
func (s *Health) SetAt(val time.Time) {
	s.At = val
}

type HealthStatus string

const (
	HealthStatusOk HealthStatus = "ok"
)

// AllValues returns all HealthStatus values.
func (HealthStatus) AllValues() []HealthStatus {
	return []HealthStatus{
		HealthStatusOk,
	}
}

// MarshalText implements encoding.TextMarshaler.
func (s HealthStatus) MarshalText() ([]byte, error) {
	switch s {
	case HealthStatusOk:
		return []byte(s), nil
	default:
		return nil, errors.Errorf("invalid value: %q", s)
	}
}

// UnmarshalText implements encoding.TextUnmarshaler.
func (s *HealthStatus) UnmarshalText(data []byte) error {
	switch HealthStatus(data) {
	case HealthStatusOk:
		*s = HealthStatusOk
		return nil
	default:
		return errors.Errorf("invalid value: %q", data)
	}
}

// NewOptString returns new OptString with value set to v.
func NewOptString(v string) OptString {
	return OptString{
		Value: v,
		Set:   true,
	}
}

// OptString is optional string.
type OptString struct {
	Value string
	Set   bool
}

// IsSet returns true if OptString was set.
func (o OptString) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptString) Reset() {
	var v string
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptString) SetTo(v string) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptString) Get() (v string, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptString) Or(d string) string {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// NewOptURI returns new OptURI with value set to v.
func NewOptURI(v url.URL) OptURI {
	return OptURI{
		Value: v,
		Set:   true,
	}
}

// OptURI is optional url.URL.
type OptURI struct {
	Value url.URL
	Set   bool
}

// IsSet returns true if OptURI was set.
func (o OptURI) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptURI) Reset() {
	var v url.URL
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptURI) SetTo(v url.URL) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptURI) Get() (v url.URL, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptURI) Or(d url.URL) url.URL {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// RFC 9457 problem details.
// Ref: #/components/schemas/Problem
type Problem struct {
	Type     OptURI    `json:"type"`
	Title    string    `json:"title"`
	Status   int       `json:"status"`
	Detail   OptString `json:"detail"`
	Instance OptURI    `json:"instance"`
}

// GetType returns the value of Type.
func (s *Problem) GetType() OptURI {
	return s.Type
}

// GetTitle returns the value of Title.
func (s *Problem) GetTitle() string {
	return s.Title
}

// GetStatus returns the value of Status.
func (s *Problem) GetStatus() int {
	return s.Status
}

// GetDetail returns the value of Detail.
func (s *Problem) GetDetail() OptString {
	return s.Detail
}

// GetInstance returns the value of Instance.
func (s *Problem) GetInstance() OptURI {
	return s.Instance
}

// SetType sets the value of Type.
func (s *Problem) SetType(val OptURI) {
	s.Type = val
}

// SetTitle sets the value of Title.
func (s *Problem) SetTitle(val string) {
	s.Title = val
}

// SetStatus sets the value of Status.
func (s *Problem) SetStatus(val int) {
	s.Status = val
}

// SetDetail sets the value of Detail.
func (s *Problem) SetDetail(val OptString) {
	s.Detail = val
}

// SetInstance sets the value of Instance.
func (s *Problem) SetInstance(val OptURI) {
	s.Instance = val
}

// ProblemStatusCode wraps Problem with StatusCode.
type ProblemStatusCode struct {
	StatusCode int
	Response   Problem
}

// GetStatusCode returns the value of StatusCode.
func (s *ProblemStatusCode) GetStatusCode() int {
	return s.StatusCode
}

// GetResponse returns the value of Response.
func (s *ProblemStatusCode) GetResponse() Problem {
	return s.Response
}

// SetStatusCode sets the value of StatusCode.
func (s *ProblemStatusCode) SetStatusCode(val int) {
	s.StatusCode = val
}

// SetResponse sets the value of Response.
func (s *ProblemStatusCode) SetResponse(val Problem) {
	s.Response = val
}